Prompt Color Shows Last Error
Had trouble with this before and forgot to return to it. The prompt
FIRST THING checks the last error ($?) variable and sets the color of
the line (-----) yellow or red, depending.

I also adjusted the > character at the end of the prompt so it's in the
"normal" forground/background colors. I might use those colors later in
the body of the prompt... but maybe not.