HUGE: Customizable!!
I added Git info ... and my eyes were opened. I re-worked the entire
thing to be customizable - you can change colors, framing characters,
even turn off pieces (some). And, you can save your preferences - even
for different users.

I'm sure it's not the best way to do some of this, and I'm also sure I'm
not done, but this is a pretty fair start!

Also, I added a feature so the Directory section indicates if a folder
is read-only.