Turn off Uptime; Add color separation tween drive and folder
Since the prompt shows a lot of info, I made one more piece turn
off-able. The Uptime section now has an UptimeOn setting. Set to false,
and no uptime display.

Since I'm using PSDrives, I've separated the drive from the folder via
colors. It's not as noticeable with a single letter (C:), but it's
obvious with the longer names.