Backed Git Status into private helper file
Somehow (and maybe this was included and removed or maybe I downloaded
the files and just put them there) but I lost the posh-git module that
was "included" with my github install. Don't know what happened; don't
care. Posh-Git is a separately maintained module anyway (see the readme
for this for the link - it's good stuff).

So, I baked git status (and only git status) into the prompt via a
helper script.

I also updated the readme file - cuz reasons.