Refactored WindowTitlePrefix
The WindowTitlePrefix was set in the profile w global scope but used in this module. Bad juju.

I've refactored it into this module (where it belongs) and made it so you can change it on the fly via the Set-WindowTitle function.

I also cleaned up a conemu function that I didn't need/use (related to the window title)

I also removed a few aliases and tidied up the spacing/format (thank you, VSCode).