IP must have a gateway
Previously, the displayed IP address was a "random" first ip found. On my laptop, that was usually a private IP for hyper-v. Now, it's still a random first IP, but it at least picks one with a default gateway (or nothing if no def gateway is defined).

Oh, and I fixed the get IP to use CIMInstance rather than WMI ... it's not PS Core compatible (finally!)